requirements.txt
torch>=2.0.0
transformers>=4.37.0
bitsandbytes>=0.41.0
Pillow>=10.0.0
accelerate>=0.25.0
safetensors>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
numpy>=1.24.0
requests>=2.31.0
